feat: code RBAC system, CRUD giao vien

+++ src/main/java/com/trungtamdaotao/model/dao/AbstractDAO.java
@@ -20,7 +20,7 @@ public abstract class AbstractDAO<T> implements IGenericDAO<T> {
     }
 
     @Override
-    public T findById(int id) {
+    public T findById(Long id) {
         try (EntityManager em = DbManager.getFactory().createEntityManager()) {
             return em.find(entityClass, id);
         }
@@ -37,14 +37,14 @@ public abstract class AbstractDAO<T> implements IGenericDAO<T> {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(Long id) {
         executeInsideTransaction(em -> {
             T entity = em.find(entityClass, id);
             if (entity != null) em.remove(entity);
         });
     }
 
-    // Một hàm bổ trợ dùng Lambda để quản lý Transaction cho gọn (SOLID)
+    // Một hàm bổ trợ dùng Lambda để quản lý Transaction cho gọn
     private void executeInsideTransaction(java.util.function.Consumer<EntityManager> action) {
         EntityManager em = DbManager.getFactory().createEntityManager();
         EntityTransaction tx = em.getTransaction();
