+++ src/main/java/com/trungtamdaotao/model/dao/IGenericDAO.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface IGenericDAO<T> {
     List<T> findAll();
-    T findById(int id);
+    T findById(Long id);
     void save(T entity);
     void update(T entity);
-    void delete(int id);
+    void delete(Long id);
 }
